LlamaIndex/requirements.txt
llama-index-core
llama-index-llms-openai
python-dotenv

